+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - Berry `tcpclientasync` class for non-blocking TCP client
 - Support for GM861 1D and 2D bar code reader (#18399)
 - Berry `re` (regex) add `match2` and optional offset
+- Support for PCA9557 8-bit I/O expander (#18632)
 
 ### Breaking Changed
 - Change command ``FileUpload`` index binary data detection from >199 to >299